requirements.txt
streamlit
pandas
google-genai
pillow
openpyxl
plotly